Abstract Trusted Publishing services

Remove unused APIs


Remove old imports


remove a line
